successfully moved taskbar functions to fightpage parent component
