+++ src/components/FightPage/TaskBar/Todo.js
@@ -1,4 +1,4 @@
-function Todo({text, todo, todos, setTodos}) {
+function Todo({todo, todos, setTodos}) {
     const deleteHandler = () => {
         setTodos(todos.filter((el) => el.id !== todo.id));
     };
